Lab4/requirements.txt
numpy==1.23.3
pandas==1.5.0
matplotlib==3.6.2
opencv-python==4.6.0.66
Pillow==9.3.0